fixing some request input on graphql

+++ pkg/cubawheeler/wallet.go
@@ -8,13 +8,13 @@ import (
 type Wallet struct {
 	ID        string        `json:"id" bson:"_id"`
 	Owner     string        `json:"owner" bson:"owner"`
-	Balance   int           `json:"balance" bson:"balance"`
+	Balance   int64         `json:"balance" bson:"balance"`
 	CreatedAt uint          `json:"-" bson:"created_at"`
 	UpdatedAt uint          `json:"updated_at" bson:"updated_at"`
 	Events    []interface{} `json:"-" bson:"events"`
 }
 
-func (w *Wallet) Deposit(amount int) {
+func (w *Wallet) Deposit(amount int64) {
 	w.Balance += amount
 	w.UpdatedAt = uint(time.Now().Unix())
 	w.Events = append(w.Events, DepositEvent{
@@ -23,7 +23,7 @@ func (w *Wallet) Deposit(amount int) {
 	})
 }
 
-func (w *Wallet) Withdraw(amount int) {
+func (w *Wallet) Withdraw(amount int64) {
 	w.Balance -= amount
 	w.UpdatedAt = uint(time.Now().Unix())
 	w.Events = append(w.Events, WithdrawEvent{
@@ -42,20 +42,20 @@ func NewWallet() *Wallet {
 }
 
 type DepositEvent struct {
-	Amount    int  `json:"amount" bson:"amount"`
-	CreatedAt uint `json:"-" bson:"created_at"`
+	Amount    int64 `json:"amount" bson:"amount"`
+	CreatedAt uint  `json:"-" bson:"created_at"`
 }
 
 type WithdrawEvent struct {
-	Amount    int  `json:"amount" bson:"amount"`
-	CreatedAt uint `json:"-" bson:"created_at"`
+	Amount    int64 `json:"amount" bson:"amount"`
+	CreatedAt uint  `json:"-" bson:"created_at"`
 }
 
 type WalletService interface {
 	Create(context.Context, string) (*Wallet, error)
-	Deposit(context.Context, string, int) (*Wallet, error)
-	Withdraw(context.Context, string, int) (*Wallet, error)
-	Transfer(context.Context, string, string, int) (*Wallet, *Wallet, error)
+	Deposit(context.Context, string, int64) (*Wallet, error)
+	Withdraw(context.Context, string, int64) (*Wallet, error)
+	Transfer(context.Context, string, string, int64) (*Wallet, *Wallet, error)
 	FindByOwner(context.Context, string) (*Wallet, error)
-	Balance(context.Context, string) (int, error)
+	Balance(context.Context, string) (int64, error)
 }
